projects/figma-search-bot/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
flask>=2.3.0
lark-oapi>=1.0.0
